Rails3: db migrate: change value of "repositories" table "type" column
The single-table inheritance mechanism failed to locate the subclass: 'child'.
This error is raised because the column 'type' is reserved for storing the class in case of inheritance.
Please rename this column if you didn't intend it to be used for storing the inheritance class
or overwrite Parent.inheritance_column to use another column for that information.
(ActiveRecord::SubclassNotFound)

--HG--
extra : rebase_source : 898ef6e07397caa0641ab51406c789b6c3098f08