Port the maths to position the menu.
The menu would always open at the position of the mouse, if that
happened to be towards the edge of the screen then it would mean having
to scroll to view it. This should make sure that it always opens within
the visible part of the view port.

This hasn't been tested since I don't have a local install to test it
with, i.e. it might have some maths mistakes.