Replace rake generate_session_store with generate_secret_token
Rails 3 uses a secret token in various places. The former cookie
secret has thus been re-purposed to a generic secret token which
is generated with the new task.

The former task also configured the cookie session store. this is
now done in application.rb and can be overridden in the
additional_environment.rb